pyproject.toml
[tool.poetry]
name = "solar"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.3"
psycopg = {extras = ["binary"], version = "^3.1.9"}
pymodbus = "^3.4.1"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
